package.json: Add npm prepare script
We need this because we are installing directly from git instead of
via npmjs.com so we need to make sure that npm builds the libraries
after installing.

The caveat is that we need to use --legacy-peer-deps when building
md-menu because of some bug in npm that causes it to prefer higher
versions when range is given (we have @angular/forms@^9.0.0...):

> npm ERR! Could not resolve dependency:
> npm ERR! peer @angular/common@"10.2.5" from @angular/forms@10.2.5
> npm ERR! node_modules/@angular/forms
> npm ERR!   peer @angular/forms@"^9.0.0 || ^10.0.0-0" from @angular/material@9.2.4
> npm ERR!   node_modules/@angular/material
> npm ERR!     peer @angular/material@">=7.0.0 <10.0.0" from the root project

Also, it seems the npm postinstall was not necessary.

See: https://github.com/npm/cli/issues/3152