Add .npmignore
If npm doesn't find an .npmignore file it will use the .gitignore
file, which usually ignores compiled files in the dist/ directory.
Because we want to be able to build this as a dependency straight
from git (instead of npmjs.com) we need to make sure npm doesn't
ignore the dist files generated by our prepare script.